#fix for first version

+++ app/BookShelf/BookShelfApi/Models/Book.cs
@@ -4,14 +4,14 @@ namespace BookShelfApi.Models
 {
     public class Book
     {
-        public int BookId { get; set; }
+        public long BookId { get; set; }
         public string? BookName { get; set; }
 
         public static Book FromRow(DataRow row)
         {
             var book = new Book();
 
-            book.BookId = row.Field<int>("BookId");
+            book.BookId = row.Field<long>("BookId");
             book.BookName = row.Field<string>("BookName");
 
             return book;
